2.4.0

+++ src/4-services/0-mos-router.js
@@ -1,15 +1,16 @@
-import ackService from "./ack-service.js";
+import ackService from "./5-ack-service.js";
 import logger from "../3-utilities/logger.js";
-import octopusService from "./octopus-service.js";
-import appProcessor from "./app-processor.js";
+import storyService from "./2-story-service.js";
+import appProcessor from "./1-app-processor.js";
 import findRoID from "../3-utilities/findRoID.js";
 import appConfig from "../3-utilities/app-config.js";
 import EventEmitter from 'events';
-import itemsService from "./items-service.js";
+import itemsService from "./3-items-service.js";
 
 const logMosIncomingMessages = appConfig.debug.showMos;
 
 class MosRouter extends EventEmitter {
+
     
     async mosMessageProcessor(msg, port) {
     
@@ -26,6 +27,8 @@ class MosRouter extends EventEmitter {
                 ackService.sendHeartbeat(port);
                 break;
             case !!msg.mos.roMetadataReplace:
+                logger(`[MOS] {${this.color("roMetadataReplace")}} are received from ${port}`);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roMetadataReplace", msg);break;}
                 await appProcessor.roMetadataReplace(msg);
                 ackService.sendAck(msg.mos.roMetadataReplace.roID);
                 break;              
@@ -39,6 +42,7 @@ class MosRouter extends EventEmitter {
                 break;
             case !!msg.mos.roCreate:
                 logger(`[MOS] {${this.color("roCreate")}} are received from ${port}`);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roCreate", msg);break;}
                 await appProcessor.roCreate(msg);
                 break;
             case !!msg.mos.roReadyToAir:
@@ -47,6 +51,7 @@ class MosRouter extends EventEmitter {
                 break;
             case !!msg.mos.roDelete:
                 logger(`[MOS] {${this.color("roDelete")}} are received from ${port}`);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roDelete", msg);break;}                
                 await appProcessor.roDelete(msg);
                 break; 
             
@@ -57,37 +62,44 @@ class MosRouter extends EventEmitter {
                 break;            
             case !!msg.mos.roStoryAppend:
                 logger(`[MOS] {${this.color("roStoryAppend")}} are received from ${port}`);
-                await octopusService.appendStory(msg);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roStoryAppend", msg);break;}
+                await storyService.appendStory(msg);
                 break; 
             case !!msg.mos.roStoryReplace:
                 logger(`[MOS] {${this.color("roStoryReplace")}} are received from ${port}`);
-                octopusService.replaceStory(msg);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roStoryReplace", msg);break;}
+                await storyService.replaceStory(msg);
                 break;             
             case !!msg.mos.roStoryDelete:
                 logger(`[MOS] {${this.color("roStoryDelete")}} are received from ${port}`);
-                await octopusService.deleteStoriesHandler(msg);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roStoryDelete", msg);break;}
+                await storyService.deleteStoriesHandler(msg);
                 break;                 
             case !!msg.mos.roStoryInsert:
                 logger(`[MOS] {${this.color("roStoryInsert")}} are received from ${port}`);
-                await octopusService.insertStory(msg);
-                //ackService.sendAck(msg.mos.roStoryInsert.roID);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roStoryInsert", msg);break;}
+                await storyService.insertStory(msg);
                 break;  
             case !!msg.mos.roStoryMoveMultiple:
                 logger(`[MOS] {${this.color("roStoryMoveMultiple ")}} are received from ${port}`);
-                octopusService.moveMultiple(msg);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roStoryMoveMultiple", msg);break;}
+                await storyService.moveMultiple(msg);
                 break;  
             
             // Items events
             case !!msg.mos.roItemInsert:
                 logger(`[MOS] {${this.color("roItemInsert")}} are received from ${port}`);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roItemInsert", msg);break;}
                 await itemsService.insertItem(msg);                    
                 break;                              
             case !!msg.mos.roItemDelete:
                 logger(`[MOS] {${this.color("roItemDelete")}} are received from ${port}`);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roItemDelete", msg);break;}
                 await itemsService.deleteItem(msg);
                 break; 
             case !!msg.mos.roItemReplace:
                 logger(`[MOS] {${this.color("roItemReplace")}} are received from ${port}`);
+                if(appProcessor.isBoot){appProcessor.enqueueDelta("roItemReplace", msg);break;}
                 await itemsService.replaceItem(msg);
                 break;                             
             // ****************** roAck handling ************************
